% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_all}
\alias{dist_all}
\title{Compute Euclidean distances between all pairs of rows
Returns numeric matrix, scaled to [0,1]}
\usage{
dist_all(X)
}
\arguments{
\item{X}{A numeric matrix}
}
\description{
Compute Euclidean distances between all pairs of rows
Returns numeric matrix, scaled to [0,1]
}
