% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dist_one}
\alias{dist_one}
\title{Compute Euclidean distances between a specified row and all other row
Returns an unscaled positive numeric vector}
\usage{
dist_one(X, x)
}
\arguments{
\item{X}{A numeric matrix}

\item{x}{a row index}
}
\description{
Compute Euclidean distances between a specified row and all other row
Returns an unscaled positive numeric vector
}
