% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albatross_functions.R
\name{man_multi}
\alias{man_multi}
\title{Calculate mid value statistic for each group in the distance matrix over a grid of cut offs}
\usage{
man_multi(dist_mat, cut_offs = seq(0, 1, length.out = 100),
  group = unique(colnames(dist_mat)), res = 3)
}
\arguments{
\item{dist_mat}{A [0, 1] bounded square distance matrix with column names matching groups.}

\item{cut_offs}{An ordered (low to high) sequence of cut_offs to calculate mid values.}

\item{group}{A character vector specifying groups of interest. Defualt is all groups.}

\item{res}{The number of digits to round mid values to. Helpful when plateaus occur.}
}
\value{
A tibble containing the following columns:
cut_offs - [0,1] cut off values input by user
mids - the mid values for each cut off
group - the group id taken from the column names of the input distance matrix
}
\description{
Calculate mid value statistic for each group in the distance matrix over a grid of cut offs
}
\examples{
sepal_dist <- iris$Sepal.Width \%>\% dist() \%>\% as.matrix()
sepal_dist <- sepal_dist/max(sepal_dist)
colnames(sepal_dist) <- iris$Species
cut_df <- man_multi(dist_mat = sepal_dist, res = 2)
ggplot(cut_df, aes(cut_offs, cut_df$mids, colour = group)) + 
geom_line()
}
