% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albatross_functions.R
\name{mid_score}
\alias{mid_score}
\title{Calculate Midpoint}
\usage{
mid_score(dist_mat, cut_off = 0, group = colnames(dist_mat)[1])
}
\arguments{
\item{dist_mat}{A [0, 1] bounded square distance matrix with column names matching groups.}

\item{cut_off}{A [0, 1] bounded scalar that specifies cutoff. Defaults to 0.}

\item{group}{A character string specifying group of interest. Default if first group in matrix.}
}
\value{
returns [0, 1] bounded midpoint
}
\description{
For a given [0,1] cutoff value, this function calculates the midpoint between within group clustering and out-of-group exclusion.
It's primary purpose is internal use for other functions.
}
\examples{
set.seed(123)
groups <- 5
n_groups <- 5
n_obs <- groups * n_groups
group_names <- rep(letters[1:groups], each = 5)
example_dist <- sample(1:100, size = 25) \%>\% 
  dist \%>\% 
as.matrix
example_dist <- example_dist/max(example_dist)
colnames(example_dist) <- group_names
mid_score(example_dist, cut_off = 0.2, group_names[1])
}
