% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/albatross_functions.R
\name{summarise_multi}
\alias{summarise_multi}
\title{Create useful summary statistics from the output of man_multi}
\usage{
summarise_multi(cut_df)
}
\arguments{
\item{cut_df}{The output of man_multi() (?man_multi() for further details)}
}
\value{
A tibble contain a row for each group id returned by man_multi():
cut_off - The miniumum cut off value of the the max mid value
plateau - The the numer cut off values that with the same max mid value
mid - The max mid value (1.0 means perfectly grouped)
mid_sum - The sum of the mid values across all considered cut_offs, which is useful for quantifying how isolated the groups are
rel_sum - The mid_sum scores, scaled by the maximum mid_sum.
}
\description{
Create useful summary statistics from the output of man_multi
}
\examples{
sepal_dist <- iris$Sepal.Width \%>\% dist() \%>\% as.matrix()
sepal_dist <- sepal_dist/max(sepal_dist)
colnames(sepal_dist) <- iris$Species
cut_df <- man_multi(dist_mat = sepal_dist, res = 2)
summarise_multi(cut_df)
}
